+++ .eslintrc.js
@@ -11,13 +11,18 @@ module.exports = {
     settings: {
         'json/sort-package-json': false,
     },
+    rules: {
+        'import/no-unresolved': 'off',
+        'import/extensions': 'off',
+        'import/no-extraneous-dependencies': 'off',
+    },
     overrides: [
         {
             files: ['**/*.ts', '**/*.tsx'],
             parser: '@typescript-eslint/parser',
             parserOptions: {
-                project: './tsconfig.json'
-            }
+                project: './tsconfig.json',
+            },
         }
-    ]
+    ],
 };