+++ packages/components/pie-modal/src/index.ts
@@ -1,10 +1,10 @@
-import { LitElement, unsafeCSS } from 'lit'; // eslint-disable-line import/no-extraneous-dependencies
+import { LitElement, unsafeCSS } from 'lit';
 import { html, unsafeStatic } from 'lit/static-html.js';
-import { property } from 'lit/decorators.js'; // eslint-disable-line import/no-extraneous-dependencies
+import { property, query } from 'lit/decorators.js'; // eslint-disable-line import/no-extraneous-dependencies
 import { RtlMixin, validPropertyValues, requiredProperty } from '@justeattakeaway/pie-webc-core';
 
 import styles from './modal.scss?inline';
-import { ModalProps, headingLevels } from './defs';
+import { ModalProps, headingLevels, ON_MODAL_CLOSE_EVENT } from './defs';
 
 // Valid values available to consumers
 export { type ModalProps, headingLevels };
@@ -23,24 +23,51 @@ export class PieModal extends RtlMixin(LitElement) {
     @validPropertyValues(componentSelector, headingLevels, 'h2')
         headingLevel: ModalProps['headingLevel'] = 'h2';
 
+    @query('dialog')
+        _dialog: HTMLDialogElement;
+
     render () {
         const {
             isOpen,
             heading,
-            headingLevel,
+            headingLevel = 'h2',
         } = this;
 
         const headingTag = unsafeStatic(headingLevel);
 
         return html`
-            <dialog class="c-modal" ?open="${isOpen}">
-                <${headingTag} class="c-modal-heading">${heading}</${headingTag}>
-                <div class="c-modal-contentWrapper">
-                    <slot></slot>
-                </div>
+            <dialog id="dialog" class="c-modal" ?open="${isOpen}">
+                <header>
+                    <${headingTag} class="c-modal-heading">${heading}</${headingTag}>
+                    <pie-icon-button
+                        @click="${this._handleCloseDialog}"
+                        variant="ghost-tertiary"
+                        class="c-modal-closeBtn"></pie-icon-button>
+                </header>
+                <article>
+                    <div class="c-modal-content">
+                        <slot></slot>
+                    </div>
+                </article>
             </dialog>
         `;
     }
+
+    _handleCloseDialog = () => {
+        this._dialog.close();
+        this._onDialogClose();
+    };
+
+    /**
+     * Dispatch `ON_MODAL_CLOSE_EVENT` event.
+     * To be used whenever we close the modal.
+     *
+     * @event
+     */
+    _onDialogClose = () => {
+        const event = new CustomEvent(ON_MODAL_CLOSE_EVENT);
+        this.dispatchEvent(event);
+    };
 
     // Renders a `CSSResult` generated from SCSS by Vite
     static styles = unsafeCSS(styles);