+++ packages/tools/pie-icons-configs/__tests__/configs.test.js
@@ -1,5 +1,5 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 import {
+    afterEach,
     describe,
     it,
     expect,
@@ -14,7 +14,13 @@ import {
     getSvgProps,
 } from '../configs';
 
+const defaultRegularIconSize = sizeToValueMap[regularIconSizeDefault];
+
 describe('validateGetLargeIconSize', () => {
+    afterEach(() => {
+        vi.restoreAllMocks();
+    });
+
     describe('when provided with valid iconSize value', () => {
         it('returns an object with the expected keys and values', () => {
             const received = validateGetLargeIconSize(largeIconSizeDefault);
@@ -67,7 +73,7 @@ describe('validateGetRegularIconSize', () => {
         it('returns an object with the expected keys and values', () => {
             const received = validateGetRegularIconSize('xs');
             const expected = {
-                iconSize: sizeToValueMap[regularIconSizeDefault],
+                iconSize: defaultRegularIconSize,
                 isValid: true,
             };
             expect(received).toEqual(expected);
@@ -76,7 +82,7 @@ describe('validateGetRegularIconSize', () => {
 
     describe('when provided with invalid iconSize value', () => {
         const expected = {
-            iconSize: sizeToValueMap[regularIconSizeDefault],
+            iconSize: defaultRegularIconSize,
             isValid: false,
         };
 
@@ -111,126 +117,129 @@ describe('validateGetRegularIconSize', () => {
 });
 
 describe('getSvgProps', () => {
-    describe('when provided with a class for regular size icon', () => {
-        const iconName = 'IconAppOrder';
-
-        describe('when provided with valid parameters', () => {
-            it('returns an object that contains a class and width and height properties', () => {
-                const received = getSvgProps('pie-icon pie-icon--app-order', null, regularIconSizeDefault, iconName);
+    const regularIconSizeName = 'IconTest';
+    describe('when provided with a', () => {
+        describe('svgClasses parameter', () => {
+            it('returns an object that contains a class attr with the expected classes', () => {
+                const received = getSvgProps('icon-test', null, null, regularIconSizeName);
 
                 expect(received).toHaveProperty('class');
-                expect(received).toHaveProperty('width');
-                expect(received).toHaveProperty('height');
-            });
-
-            it('has expected width and height values', () => {
-                const received = getSvgProps('pie-icon pie-icon--app-order', null, regularIconSizeDefault, iconName);
-                const expectedRegularIconSizeValue = sizeToValueMap[regularIconSizeDefault];
-
-                expect(received.width).toEqual(expectedRegularIconSizeValue);
-                expect(received.height).toEqual(expectedRegularIconSizeValue);
-            });
-
-            describe('when provided with a staticClasses parameter', () => {
-                it('returns an object with the expected classes', () => {
-                    const received = getSvgProps('pie-icon pie-icon--app-order', 'FOO BAR', regularIconSizeDefault, iconName);
-
-                    expect(received.class).toEqual(expect.stringContaining('pie-icon'));
-                    expect(received.class).toEqual(expect.stringContaining('pie-icon--app-order'));
-                    expect(received.class).toEqual(expect.stringContaining('FOO BAR'));
-                });
+                expect(received.class).toEqual('icon-test');
             });
         });
 
-        describe('when provided with invalid parameters', () => {
-            it('returns an object with the expected classes', () => {
-                const received = getSvgProps('pie-icon pie-icon--app-order', 'FOO', null, iconName);
+        describe('staticClasses parameter', () => {
+            it('returns an object that contains a class attr with the expected classes', () => {
+                const received = getSvgProps('icon-test', 'FOO BAR', null, regularIconSizeName);
 
-                expect(received.class).toEqual(expect.stringContaining('pie-icon'));
-                expect(received.class).toEqual(expect.stringContaining('pie-icon--app-order'));
+                expect(received).toHaveProperty('class');
+                expect(received.class).toEqual('icon-test FOO BAR');
             });
+        });
 
-            it('outputs a console error', () => {
-                const spy = vi.spyOn(console, 'error').mockImplementation(() => {});
+        describe('valid iconSizeValue parameter', () => {
+            it('returns an object that contains a width and height attr with the correct values', () => {
+                const received = getSvgProps('icon-test', null, regularIconSizeDefault, regularIconSizeName);
+                const expectedRegularIconSizeValue = sizeToValueMap[regularIconSizeDefault];
 
-                getSvgProps('pie-icon pie-icon--app-order', 'FOO', null, iconName);
-
-                expect(spy).toHaveBeenCalled();
-                expect(spy).toHaveBeenCalledWith(expect.stringContaining('Invalid prop "iconSize" value'));
-                expect(spy).toHaveBeenCalledWith(expect.stringContaining(iconName));
-                spy.mockRestore();
+                expect(received).toHaveProperty('width');
+                expect(received).toHaveProperty('height');
+                expect(received.width).toEqual(expectedRegularIconSizeValue);
+                expect(received.height).toEqual(expectedRegularIconSizeValue);
             });
         });
     });
 
-    describe('when provided with a class for large size icon', () => {
-        describe('when provided with valid parameters', () => {
-            it('returns an object that contains a class and width and height properties', () => {
-                const received = getSvgProps('pie-icon pie-icon--app-order-large', null, largeIconSizeDefault);
+    describe('when provided with invalid iconSizeValue parameter', () => {
+        const invalidIconSizeValue = 'n';
 
-                expect(received).toHaveProperty('class');
+        it('returns an object with default width and height properties', () => {
+            vi.spyOn(console, 'error').mockImplementation(vi.fn());
+            const received = getSvgProps('icon-test', null, invalidIconSizeValue, regularIconSizeName);
+
+            expect(received.width).toEqual(defaultRegularIconSize);
+            expect(received.height).toEqual(defaultRegularIconSize);
+        });
+
+        it('outputs a console error', () => {
+            const spy = vi.spyOn(console, 'error').mockImplementation(vi.fn());
+            getSvgProps('icon-test', null, invalidIconSizeValue, regularIconSizeName);
+
+            expect(spy).toHaveBeenCalled();
+            expect(spy).toHaveBeenCalledWith(expect.stringContaining('Invalid prop "iconSize" value'));
+            expect(spy).toHaveBeenCalledWith(expect.stringContaining(regularIconSizeName));
+        });
+    });
+
+    describe('when iconSizeValue is not provided', () => {
+        it('returns an object with default icon width and height properties', () => {
+            const received = getSvgProps('icon-test', null, null, regularIconSizeName);
+
+            expect(received.width).toEqual(defaultRegularIconSize);
+            expect(received.height).toEqual(defaultRegularIconSize);
+        });
+        it('does not output a console error', () => {
+            const spy = vi.spyOn(console, 'error').mockImplementation(vi.fn());
+
+            expect(spy).not.toHaveBeenCalled();
+        });
+    });
+
+    describe('when svgClasses parameter contains -large in class name', () => {
+        const largeIconSizeName = 'IconTestLarge';
+        describe('when provided with valid iconSizeValue parameter', () => {
+            it('returns an object that contains a width and height attr with the correct values', () => {
+                const validLargeIconSize = 48;
+                const received = getSvgProps('icon-test-large', null, validLargeIconSize, largeIconSizeName);
+
                 expect(received).toHaveProperty('width');
                 expect(received).toHaveProperty('height');
-            });
-
-            describe('when provided with a staticClasses parameter', () => {
-                it('returns an object with the expected classes', () => {
-                    const received = getSvgProps('pie-icon pie-icon--app-order-large', 'FOO BAR', largeIconSizeDefault);
-
-                    expect(received.class).toEqual(expect.stringContaining('pie-icon'));
-                    expect(received.class).toEqual(expect.stringContaining('pie-icon--app-order'));
-                    expect(received.class).toEqual(expect.stringContaining('FOO BAR'));
-                });
+                expect(received.width).toEqual(validLargeIconSize);
+                expect(received.height).toEqual(validLargeIconSize);
             });
         });
 
-        describe('when provided with invalid parameters', () => {
-            describe('when provided with an invalid iconSize', () => {
-                const iconName = 'IconAppOrder';
-                const errorMock = vi.fn();
-                console.error = errorMock;
+        describe('when provided with invalid iconSizeValue parameter', () => {
+            const invalidSize = 49;
 
-                it('returns an object with width and height properties', () => {
-                    const received = getSvgProps('pie-icon pie-icon--app-order-large', null, null, iconName);
+            it('returns an object with the default width and height', () => {
+                vi.spyOn(console, 'error').mockImplementation(vi.fn());
+                const received = getSvgProps('icon-test-large', null, invalidSize, largeIconSizeName);
 
-                    expect(received).toHaveProperty('class');
-                    expect(received).toHaveProperty('width');
-                    expect(received).toHaveProperty('height');
-                });
+                expect(received.width).toEqual(largeIconSizeDefault);
+                expect(received.height).toEqual(largeIconSizeDefault);
+            });
 
-                it('returns an object with the default width and height', () => {
-                    const received = getSvgProps('pie-icon pie-icon--app-order-large', null, null, iconName);
+            it('returns an object with the expected classes', () => {
+                vi.spyOn(console, 'error').mockImplementation(vi.fn());
+                const received = getSvgProps('icon-test-large', null, invalidSize, largeIconSizeName);
 
-                    expect(received.width).toEqual(largeIconSizeDefault);
-                    expect(received.height).toEqual(largeIconSizeDefault);
-                });
+                expect(received).toHaveProperty('class');
+                expect(received.class).toEqual('icon-test-large');
+            });
 
-                it('returns an object with the expected classes', () => {
-                    const received = getSvgProps('pie-icon pie-icon--app-order-large', null, null, iconName);
+            it('outputs a console error', () => {
+                const spy = vi.spyOn(console, 'error').mockImplementation(vi.fn());
+                getSvgProps('icon-test-large', null, invalidSize, largeIconSizeName);
 
-                    expect(received.class).toEqual(expect.stringContaining('pie-icon'));
-                    expect(received.class).toEqual(expect.stringContaining('pie-icon--app-order-large'));
-                });
+                expect(spy).toHaveBeenCalled();
+                expect(spy).toHaveBeenCalledWith(expect.stringContaining('Invalid prop "iconSize" value'));
+                expect(spy).toHaveBeenCalledWith(expect.stringContaining(largeIconSizeName));
+            });
+        });
 
-                it('returns an object with valid width and height properties', () => {
-                    const received = getSvgProps('pie-icon pie-icon--app-order-large', null, 49, iconName);
-                    const expected = 48;
-                    expect(received.width).toEqual(expected);
-                    expect(received.height).toEqual(expected);
-                });
+        describe('when iconSize is not provided', () => {
+            it('returns an object with default icon width and height properties', () => {
+                const received = getSvgProps('test-large', null, null, largeIconSizeName);
 
-                it('outputs a console error', () => {
-                    const spy = vi.spyOn(console, 'error').mockImplementation(vi.fn());
+                expect(received.width).toEqual(largeIconSizeDefault);
+                expect(received.height).toEqual(largeIconSizeDefault);
+            });
+            it('does not output a console error', () => {
+                const spy = vi.spyOn(console, 'error').mockImplementation(vi.fn());
 
-                    getSvgProps('pie-icon pie-icon--app-order-large', 'FOO', null, iconName);
-
-                    expect(spy).toHaveBeenCalled();
-                    expect(spy).toHaveBeenCalledWith(expect.stringContaining('Invalid prop "iconSize" value'));
-                    expect(spy).toHaveBeenCalledWith(expect.stringContaining(iconName));
-                    spy.mockRestore();
-                });
+                expect(spy).not.toHaveBeenCalled();
             });
         });
     });
 });
-