+++ apps/examples/wc-next10/CHANGELOG.md
@@ -1,5 +1,12 @@
 # wc-next10
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [[`16c8af6e6`](https://github.com/justeattakeaway/pie/commit/16c8af6e6c1d7206d27034b08da117532a161ac3)]:
+  - @justeattakeaway/pie-button@0.16.0
+
 ## 0.3.1
 
 ### Patch Changes