+++ packages/components/pie-icon-button/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @justeattakeaway/pie-icon-button
 
+## 0.6.0
+
+### Minor Changes
+
+- [Added] - `size` prop added to pie-icon-button ([#557](https://github.com/justeattakeaway/pie/pull/557)) by [@ashleynolan](https://github.com/ashleynolan)
+
+## 0.5.0
+
+### Minor Changes
+
+- [Updated] - enum types to string union types to string union types ([#508](https://github.com/justeattakeaway/pie/pull/508)) by [@FayeCarter](https://github.com/FayeCarter)
+
 ## 0.4.0
 
 ### Minor Changes