+++ packages/tools/pie-icons-configs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @justeattakeaway/pie-icons-configs
 
+## 3.6.0
+
+### Patch Changes
+
+- [Fixed]: removes a console error if iconSize prop is not passed ([#556](https://github.com/justeattakeaway/pie/pull/556)) by [@dandel10n](https://github.com/dandel10n)
+
+  [Changed]: in case of invalid `iconSize` being passed to large icons now the size defaults to the `largeIconSizeDefault`
+
+## 3.5.1
+
+### Patch Changes
+
+- [Changed]: removes class attr application for react icons ([#543](https://github.com/justeattakeaway/pie/pull/543)) by [@dandel10n](https://github.com/dandel10n)
+
 ## 3.5.0
 
 ### Minor Changes